backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.0
colorama==0.4.6
fastapi==0.115.12
h11==0.16.0
idna==3.10
pydantic==2.11.4
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.31
psycopg2==2.9.10
bcrypt==4.0.1
websockets==11.0.3
email-validator==2.1.0.post1
pytest==7.4.3
httpx==0.25.1
requests==2.31.0
